requirements.txt
fastapi==0.115.0
uvicorn==0.30.0
smartapi-python==1.4.2
websocket-client==1.8.0
easyocr==1.7.2
python-dotenv==1.0.0
sqlalchemy==2.0.36
aiosqlite==0.20.0
apscheduler==3.10.4
pyotp==2.9.0
pillow==10.4.0
requests==2.32.3
python-multipart==0.0.12
jinja2==3.1.4
httpx==0.27.0
pandas==2.2.3
pytz==2024.2
loguru==0.7.2
torch==2.2.2
torchvision==0.17.2
